Add needed Components and organize files
