+++ src/Components/App/App.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './App.css';
+
+export const App = () => {
+  return (
+    <main>
+
+    </main>
+  )
+}
